Moved to DateTimeUtils, spotless fixed, removed temporary fix
